feat(netstat): fold sockwho into netstat.attribution; dedupe socket walk; bound hardware WMI
Consolidates the standalone sockwho plugin into netstat as a new
`attribution` action (proto|local_addr|local_port|remote_addr|remote_port|
state|pid|process_name|process_path), then retires the sockwho plugin
entirely — meson, installer, content definitions, server registry,
capability catalogue, result_parsing schema, UAT/puppeteer suites, and docs
all updated. netstat's and tar's macOS libproc socket walks now both
consume the shared agents/shared/macos_socket_walk.hpp helper instead of
inline copies (#3403). hardware plugin's private WMI enumerator no longer
calls Next(WBEM_INFINITE, ...); it now rides the shared bounded
agents/shared/wmi_bounded.hpp helper across all 8 of its WMI call sites
(#3404).

Mechanism stays native throughout: GetExtTcpTable(+QueryFullProcessImageNameW)
on Windows, /proc/net + /proc/[pid]/{comm,exe,fd} on Linux, libproc on macOS
— no new spawns.
