+++ Synthetic-UAT-Puppeteer.js
@@ -210,7 +210,7 @@ const COMMANDS = [
 
   // ─── Process Fetch / Socket Who ───
   ['procfetch procfetch',   'Process fetch with hashes', 'any'],
-  ['sockwho sockwho_list',  'Socket-to-process map',     'any'],
+  ['netstat attribution',   'Socket-to-process map',     'any'],
 
   // ─── Diagnostics ───
   ['diagnostics log_level',       'Agent log level',       'rows'],
