+++ agents/plugins/hardware/meson.build
@@ -24,7 +24,7 @@ hardware_plugin_lib = shared_library(
   'hardware',
   files('src/hardware_plugin.cpp'),
   name_prefix: '',
-  include_directories: include_directories('../../shared'),  # win_str.hpp (#1681, _WIN32-gated)
+  include_directories: include_directories('../../shared'),  # wmi_bounded.hpp (#3404, _WIN32-gated)
   dependencies: hardware_deps,
   install: true,
   install_dir: get_option('libdir') / 'yuzu' / 'plugins',
