+++ agents/plugins/netstat/meson.build
@@ -9,6 +9,8 @@ shared_library(
   'netstat',
   files('src/netstat_plugin.cpp'),
   name_prefix: '',
+  # macos_socket_walk.hpp (#3403 dedupe, darwin) + win_str.hpp (#1681, windows attribution)
+  include_directories: include_directories('../../shared'),
   dependencies: netstat_deps,
   install: true,
   install_dir: get_option('libdir') / 'yuzu' / 'plugins',
