+++ agents/plugins/netstat/src/netstat_plugin.cpp
@@ -5,14 +5,26 @@
  *   "netstat_list" — Enumerates active TCP/UDP connections and listening
  *                    sockets on the host, returning protocol, addresses,
  *                    ports, state, and owning PID.
+ *   "attribution"  — Same enumeration, plus the owning process's name and
+ *                    executable path (formerly the standalone `sockwho`
+ *                    plugin, folded in here — #3403/roadmap D2). Emits
+ *                    netstat_list's 7 columns as a prefix with
+ *                    process_name/process_path appended (9 total); pid
+ *                    stays in its netstat_list position (6th field after
+ *                    proto), NOT sockwho's old 1st-column layout.
  *
  * Output is pipe-delimited, one connection per line via write_output():
- *   proto|local_addr|local_port|remote_addr|remote_port|state|pid
+ *   netstat_list: proto|local_addr|local_port|remote_addr|remote_port|state|pid
+ *   attribution:  proto|local_addr|local_port|remote_addr|remote_port|state|pid|process_name|process_path
  *
  * Platform support:
  *   Linux   — /proc/net/{tcp,tcp6,udp,udp6} + /proc/[pid]/fd inode mapping
- *   macOS   — libproc (proc_listallpids, proc_pidinfo, proc_pidfdinfo)
+ *             (attribution additionally reads /proc/[pid]/{comm,exe})
+ *   macOS   — libproc, via the shared agents/shared/macos_socket_walk.hpp
+ *             walk (also used by network_diag/ioc; #3403 dedupe)
  *   Windows — IP Helper API (GetExtendedTcpTable, GetExtendedUdpTable)
+ *             (attribution additionally resolves owning process via
+ *             QueryFullProcessImageNameW, cached per PID)
  */
 
 #include <yuzu/plugin.hpp>
@@ -35,8 +47,8 @@
 #include <unistd.h>
 #elif defined(__APPLE__)
 #include <arpa/inet.h>
-#include <libproc.h>
-#include <sys/proc_info.h>
+
+#include <macos_socket_walk.hpp> // shared libproc socket walk (#3403)
 #elif defined(_WIN32)
 #ifndef WIN32_LEAN_AND_MEAN
 #define WIN32_LEAN_AND_MEAN
@@ -48,9 +60,25 @@
 #include <ws2tcpip.h>
 #include <windows.h>
 #include <iphlpapi.h>
+#include <win_str.hpp> // shared yuzu::win wide<->UTF-8 helpers (#1681)
 #endif
 
 namespace {
+
+// Escape '|' in a field that may contain arbitrary text (process name/path)
+// so it can't be mistaken for a column separator downstream. Ported from the
+// retired sockwho_plugin.cpp.
+std::string escape_pipes(std::string_view sv) {
+    std::string out;
+    out.reserve(sv.size());
+    for (char c : sv) {
+        if (c == '|')
+            out += "\\|";
+        else
+            out += c;
+    }
+    return out;
+}
 
 // -- Linux implementation -----------------------------------------------------
 #ifdef __linux__
@@ -231,146 +259,178 @@ void enumerate_and_stream(yuzu::CommandContext& ctx) {
     parse_proc_net_file("/proc/net/udp6", "udp6", inode_map, ctx, false, true);
 }
 
-// -- macOS implementation -----------------------------------------------------
-#elif defined(__APPLE__)
+// -- attribution: /proc inode→pid + pid→process enrichment (ex-sockwho) -----
 
-constexpr std::string_view tcp_state_str_mac(int st) noexcept {
-    // TSI_S_* constants from <netinet/tcp_fsm.h> (included via sys/proc_info.h)
-    switch (st) {
-    case 0:
-        return "CLOSED";
-    case 1:
-        return "LISTEN";
-    case 2:
-        return "SYN_SENT";
-    case 3:
-        return "SYN_RECV";
-    case 4:
-        return "ESTABLISHED";
-    case 5:
-        return "CLOSE_WAIT";
-    case 6:
-        return "FIN_WAIT1";
-    case 7:
-        return "CLOSING";
-    case 8:
-        return "LAST_ACK";
-    case 9:
-        return "FIN_WAIT2";
-    case 10:
-        return "TIME_WAIT";
-    default:
-        return "UNKNOWN";
+struct ProcInfo {
+    std::string name;
+    std::string path;
+};
+
+// Build two maps in a single /proc scan: inode → PID (socket resolution) and
+// PID → ProcInfo (process name/path). Ported from sockwho_plugin.cpp's
+// build_maps() — sockwho is retired, this is now netstat's own attribution
+// enrichment path.
+void build_socket_and_proc_maps(std::unordered_map<uint64_t, int>& inode_map,
+                                std::unordered_map<int, ProcInfo>& proc_map) {
+    DIR* proc_dir = opendir("/proc");
+    if (!proc_dir)
+        return;
+
+    struct dirent* proc_entry = nullptr;
+    while ((proc_entry = readdir(proc_dir)) != nullptr) {
+        int pid = 0;
+        [[maybe_unused]] auto [ptr, ec] = std::from_chars(proc_entry->d_name,
+                                         proc_entry->d_name + std::strlen(proc_entry->d_name), pid);
+        if (ec != std::errc{} || pid <= 0)
+            continue;
+
+        std::string proc_path = std::format("/proc/{}", pid);
+
+        ProcInfo info;
+        {
+            std::ifstream comm_f(proc_path + "/comm");
+            std::getline(comm_f, info.name);
+        }
+        {
+            char link_buf[4096];
+            ssize_t len =
+                readlink((proc_path + "/exe").c_str(), link_buf, sizeof(link_buf) - 1);
+            if (len > 0)
+                info.path.assign(link_buf, static_cast<size_t>(len));
+        }
+        if (!info.name.empty())
+            proc_map.emplace(pid, std::move(info));
+
+        std::string fd_path = proc_path + "/fd";
+        DIR* fd_dir = opendir(fd_path.c_str());
+        if (!fd_dir)
+            continue;
+
+        char link_buf[128];
+        struct dirent* fd_entry = nullptr;
+        while ((fd_entry = readdir(fd_dir)) != nullptr) {
+            if (fd_entry->d_name[0] == '.')
+                continue;
+            std::string link_path = std::format("{}/{}", fd_path, fd_entry->d_name);
+            ssize_t len = readlink(link_path.c_str(), link_buf, sizeof(link_buf) - 1);
+            if (len <= 0)
+                continue;
+            link_buf[len] = '\0';
+
+            std::string_view sv(link_buf, static_cast<size_t>(len));
+            if (!sv.starts_with("socket:["))
+                continue;
+            auto inode_sv = sv.substr(8, sv.size() - 9);
+            uint64_t inode = 0;
+            std::from_chars(inode_sv.data(), inode_sv.data() + inode_sv.size(), inode);
+            if (inode > 0)
+                inode_map.emplace(inode, pid);
+        }
+        closedir(fd_dir);
+    }
+    closedir(proc_dir);
+}
+
+void parse_proc_net_file_attributed(const char* path, std::string_view proto,
+                                    const std::unordered_map<uint64_t, int>& inode_map,
+                                    const std::unordered_map<int, ProcInfo>& proc_map,
+                                    yuzu::CommandContext& ctx, bool is_tcp, bool is_ipv6) {
+    std::ifstream f(path);
+    if (!f)
+        return;
+
+    std::string line;
+    std::getline(f, line); // skip header
+
+    while (std::getline(f, line)) {
+        std::istringstream iss(line);
+        std::string sl, local, remote, state_hex;
+        if (!(iss >> sl >> local >> remote >> state_hex))
+            continue;
+
+        auto colon = local.rfind(':');
+        if (colon == std::string::npos)
+            continue;
+        std::string local_addr = is_ipv6 ? parse_ipv6(std::string_view(local).substr(0, colon))
+                                         : parse_ipv4(std::string_view(local).substr(0, colon));
+        uint16_t local_port = parse_hex_port(std::string_view(local).substr(colon + 1));
+
+        auto rcolon = remote.rfind(':');
+        if (rcolon == std::string::npos)
+            continue;
+        std::string remote_addr = is_ipv6 ? parse_ipv6(std::string_view(remote).substr(0, rcolon))
+                                          : parse_ipv4(std::string_view(remote).substr(0, rcolon));
+        uint16_t remote_port = parse_hex_port(std::string_view(remote).substr(rcolon + 1));
+
+        int state_val = 0;
+        std::from_chars(state_hex.data(), state_hex.data() + state_hex.size(), state_val, 16);
+        std::string_view state = is_tcp ? tcp_state_str(state_val) : std::string_view{};
+
+        std::string tok;
+        for (int i = 0; i < 5 && (iss >> tok); ++i) {}
+        uint64_t inode = 0;
+        if (iss >> inode) { /* got inode */
+        }
+
+        int pid = -1;
+        std::string_view pname, ppath;
+        if (inode > 0) {
+            auto it = inode_map.find(inode);
+            if (it != inode_map.end()) {
+                pid = it->second;
+                auto pit = proc_map.find(pid);
+                if (pit != proc_map.end()) {
+                    pname = pit->second.name;
+                    ppath = pit->second.path;
+                }
+            }
+        }
+
+        ctx.write_output(std::format("{}|{}|{}|{}|{}|{}|{}|{}|{}", proto, local_addr, local_port,
+                                     remote_addr, remote_port, state, pid, escape_pipes(pname),
+                                     escape_pipes(ppath)));
     }
 }
 
-std::string format_addr4(const struct in_addr& addr) {
-    char buf[INET_ADDRSTRLEN]{};
-    inet_ntop(AF_INET, &addr, buf, sizeof(buf));
-    return buf;
+void enumerate_and_stream_attribution(yuzu::CommandContext& ctx) {
+    std::unordered_map<uint64_t, int> inode_map;
+    std::unordered_map<int, ProcInfo> proc_map;
+    build_socket_and_proc_maps(inode_map, proc_map);
+
+    parse_proc_net_file_attributed("/proc/net/tcp", "tcp", inode_map, proc_map, ctx, true, false);
+    parse_proc_net_file_attributed("/proc/net/tcp6", "tcp6", inode_map, proc_map, ctx, true, true);
+    parse_proc_net_file_attributed("/proc/net/udp", "udp", inode_map, proc_map, ctx, false, false);
+    parse_proc_net_file_attributed("/proc/net/udp6", "udp6", inode_map, proc_map, ctx, false, true);
 }
 
-std::string format_addr6(const struct in6_addr& addr) {
-    char buf[INET6_ADDRSTRLEN]{};
-    inet_ntop(AF_INET6, &addr, buf, sizeof(buf));
-    return buf;
-}
+// -- macOS implementation -----------------------------------------------------
+#elif defined(__APPLE__)
 
+// netstat_list's own libproc walk now rides the shared header (#3403 dedupe)
+// instead of an inline copy — byte-parity with the pre-migration output
+// (dedup=true matches the removed inline walk's fork-dedup behaviour).
 void enumerate_and_stream(yuzu::CommandContext& ctx) {
-    // Get list of all PIDs
-    int pid_count = proc_listallpids(nullptr, 0);
-    if (pid_count <= 0)
-        return;
+    for (const auto& s : yuzu::shared::walk_sockets(/*dedup=*/true)) {
+        ctx.write_output(std::format("{}|{}|{}|{}|{}|{}|{}", s.proto, s.local_addr, s.local_port,
+                                     s.remote_addr, s.remote_port, s.state,
+                                     static_cast<int>(s.pid)));
+    }
+}
 
-    std::vector<pid_t> pids(static_cast<size_t>(pid_count) * 2); // over-allocate
-    pid_count = proc_listallpids(pids.data(), static_cast<int>(pids.size() * sizeof(pid_t)));
-    if (pid_count <= 0)
-        return;
-    pids.resize(static_cast<size_t>(pid_count));
-
-    // Track seen connections to avoid duplicates (key: proto+local+remote)
-    std::unordered_map<std::string, bool> seen;
-
-    for (pid_t pid : pids) {
-        // Get file descriptor list for this process
-        int buf_size = proc_pidinfo(pid, PROC_PIDLISTFDS, 0, nullptr, 0);
-        if (buf_size <= 0)
-            continue;
-
-        auto fd_count = static_cast<size_t>(buf_size) / sizeof(struct proc_fdinfo);
-        std::vector<struct proc_fdinfo> fds(fd_count);
-        int actual = proc_pidinfo(pid, PROC_PIDLISTFDS, 0, fds.data(),
-                                  static_cast<int>(fds.size() * sizeof(struct proc_fdinfo)));
-        if (actual <= 0)
-            continue;
-        fd_count = static_cast<size_t>(actual) / sizeof(struct proc_fdinfo);
-
-        for (size_t i = 0; i < fd_count; ++i) {
-            if (fds[i].proc_fdtype != PROX_FDTYPE_SOCKET)
-                continue;
-
-            struct socket_fdinfo si{};
-            int si_size =
-                proc_pidfdinfo(pid, fds[i].proc_fd, PROC_PIDFDSOCKETINFO, &si, sizeof(si));
-            if (si_size < static_cast<int>(sizeof(si)))
-                continue;
-
-            int family = si.psi.soi_family;
-            if (family != AF_INET && family != AF_INET6)
-                continue;
-
-            int kind = si.psi.soi_kind;
-            bool is_tcp = (kind == SOCKINFO_TCP);
-            bool is_udp = (kind == SOCKINFO_IN);
-            if (!is_tcp && !is_udp)
-                continue;
-
-            std::string proto;
-            std::string local_addr, remote_addr;
-            uint16_t local_port = 0, remote_port = 0;
-            std::string_view state;
-
-            if (is_tcp) {
-                auto& tcp = si.psi.soi_proto.pri_tcp;
-                state = tcp_state_str_mac(tcp.tcpsi_state);
-
-                if (family == AF_INET) {
-                    proto = "tcp";
-                    local_addr = format_addr4(tcp.tcpsi_ini.insi_laddr.ina_46.i46a_addr4);
-                    remote_addr = format_addr4(tcp.tcpsi_ini.insi_faddr.ina_46.i46a_addr4);
-                } else {
-                    proto = "tcp6";
-                    local_addr = format_addr6(tcp.tcpsi_ini.insi_laddr.ina_6);
-                    remote_addr = format_addr6(tcp.tcpsi_ini.insi_faddr.ina_6);
-                }
-                local_port = ntohs(static_cast<uint16_t>(tcp.tcpsi_ini.insi_lport));
-                remote_port = ntohs(static_cast<uint16_t>(tcp.tcpsi_ini.insi_fport));
-            } else {
-                auto& inp = si.psi.soi_proto.pri_in;
-
-                if (family == AF_INET) {
-                    proto = "udp";
-                    local_addr = format_addr4(inp.insi_laddr.ina_46.i46a_addr4);
-                    remote_addr = "*";
-                } else {
-                    proto = "udp6";
-                    local_addr = format_addr6(inp.insi_laddr.ina_6);
-                    remote_addr = "*";
-                }
-                local_port = ntohs(static_cast<uint16_t>(inp.insi_lport));
-                remote_port = 0;
-            }
-
-            // Deduplicate — same socket may appear in multiple PIDs (fork)
-            auto key = std::format("{}:{}:{}:{}:{}", proto, local_addr, local_port, remote_addr,
-                                   remote_port);
-            if (seen.contains(key))
-                continue;
-            seen.emplace(std::move(key), true);
-
-            ctx.write_output(std::format("{}|{}|{}|{}|{}|{}|{}", proto, local_addr, local_port,
-                                         remote_addr, remote_port, state, static_cast<int>(pid)));
+// attribution: the same shared walk, plus resolve_proc_name_path() per row
+// (sourced from sockwho_plugin.cpp originally — see macos_socket_walk.hpp's
+// file comment; sockwho itself is retired).
+void enumerate_and_stream_attribution(yuzu::CommandContext& ctx) {
+    for (const auto& s : yuzu::shared::walk_sockets(/*dedup=*/true)) {
+        std::string pname, ppath;
+        if (auto resolved = yuzu::shared::resolve_proc_name_path(s.pid)) {
+            pname = std::move(resolved->first);
+            ppath = std::move(resolved->second);
         }
+        ctx.write_output(std::format("{}|{}|{}|{}|{}|{}|{}|{}|{}", s.proto, s.local_addr,
+                                     s.local_port, s.remote_addr, s.remote_port, s.state,
+                                     static_cast<int>(s.pid), escape_pipes(pname),
+                                     escape_pipes(ppath)));
     }
 }
 
@@ -543,14 +603,180 @@ void enumerate_and_stream(yuzu::CommandContext& ctx) {
     emit_udp6(ctx);
 }
 
+// -- attribution: QueryFullProcessImageNameW enrichment (ex-sockwho) --------
+
+struct ProcInfo {
+    std::string name;
+    std::string path;
+};
+
+using yuzu::win::from_wide;
+
+ProcInfo get_proc_info(DWORD pid) {
+    ProcInfo info;
+    HANDLE proc = OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION, FALSE, pid);
+    if (!proc)
+        return info;
+
+    wchar_t path_buf[MAX_PATH];
+    DWORD path_len = MAX_PATH;
+    if (QueryFullProcessImageNameW(proc, 0, path_buf, &path_len)) {
+        info.path = from_wide(path_buf);
+        auto slash = info.path.rfind('\\');
+        if (slash != std::string::npos)
+            info.name = info.path.substr(slash + 1);
+        else
+            info.name = info.path;
+    }
+    CloseHandle(proc);
+    return info;
+}
+
+const ProcInfo& lookup_proc(DWORD pid, std::unordered_map<DWORD, ProcInfo>& cache) {
+    auto it = cache.find(pid);
+    if (it != cache.end())
+        return it->second;
+    auto [inserted, ok] = cache.emplace(pid, get_proc_info(pid));
+    return inserted->second;
+}
+
+void emit_tcp4_attributed(yuzu::CommandContext& ctx, std::unordered_map<DWORD, ProcInfo>& cache) {
+    DWORD size = 0;
+    GetExtendedTcpTable(nullptr, &size, FALSE, AF_INET, TCP_TABLE_OWNER_PID_ALL, 0);
+    if (size == 0)
+        return;
+
+    std::vector<BYTE> buf(size);
+    for (int attempt = 0; attempt < 3; ++attempt) {
+        DWORD ret =
+            GetExtendedTcpTable(buf.data(), &size, FALSE, AF_INET, TCP_TABLE_OWNER_PID_ALL, 0);
+        if (ret == NO_ERROR)
+            break;
+        if (ret == ERROR_INSUFFICIENT_BUFFER) {
+            buf.resize(size);
+            continue;
+        }
+        return;
+    }
+
+    auto* table = reinterpret_cast<MIB_TCPTABLE_OWNER_PID*>(buf.data());
+    for (DWORD i = 0; i < table->dwNumEntries; ++i) {
+        auto& row = table->table[i];
+        auto& proc = lookup_proc(row.dwOwningPid, cache);
+        ctx.write_output(std::format(
+            "tcp|{}|{}|{}|{}|{}|{}|{}|{}", format_addr4(row.dwLocalAddr),
+            ntohs(static_cast<u_short>(row.dwLocalPort)), format_addr4(row.dwRemoteAddr),
+            ntohs(static_cast<u_short>(row.dwRemotePort)), tcp_state_str_win(row.dwState),
+            static_cast<int>(row.dwOwningPid), escape_pipes(proc.name), escape_pipes(proc.path)));
+    }
+}
+
+void emit_tcp6_attributed(yuzu::CommandContext& ctx, std::unordered_map<DWORD, ProcInfo>& cache) {
+    DWORD size = 0;
+    GetExtendedTcpTable(nullptr, &size, FALSE, AF_INET6, TCP_TABLE_OWNER_PID_ALL, 0);
+    if (size == 0)
+        return;
+
+    std::vector<BYTE> buf(size);
+    for (int attempt = 0; attempt < 3; ++attempt) {
+        DWORD ret =
+            GetExtendedTcpTable(buf.data(), &size, FALSE, AF_INET6, TCP_TABLE_OWNER_PID_ALL, 0);
+        if (ret == NO_ERROR)
+            break;
+        if (ret == ERROR_INSUFFICIENT_BUFFER) {
+            buf.resize(size);
+            continue;
+        }
+        return;
+    }
+
+    auto* table = reinterpret_cast<MIB_TCP6TABLE_OWNER_PID*>(buf.data());
+    for (DWORD i = 0; i < table->dwNumEntries; ++i) {
+        auto& row = table->table[i];
+        auto& proc = lookup_proc(row.dwOwningPid, cache);
+        ctx.write_output(std::format(
+            "tcp6|{}|{}|{}|{}|{}|{}|{}|{}", format_addr6(row.ucLocalAddr),
+            ntohs(static_cast<u_short>(row.dwLocalPort)), format_addr6(row.ucRemoteAddr),
+            ntohs(static_cast<u_short>(row.dwRemotePort)), tcp_state_str_win(row.dwState),
+            static_cast<int>(row.dwOwningPid), escape_pipes(proc.name), escape_pipes(proc.path)));
+    }
+}
+
+void emit_udp4_attributed(yuzu::CommandContext& ctx, std::unordered_map<DWORD, ProcInfo>& cache) {
+    DWORD size = 0;
+    GetExtendedUdpTable(nullptr, &size, FALSE, AF_INET, UDP_TABLE_OWNER_PID, 0);
+    if (size == 0)
+        return;
+
+    std::vector<BYTE> buf(size);
+    for (int attempt = 0; attempt < 3; ++attempt) {
+        DWORD ret = GetExtendedUdpTable(buf.data(), &size, FALSE, AF_INET, UDP_TABLE_OWNER_PID, 0);
+        if (ret == NO_ERROR)
+            break;
+        if (ret == ERROR_INSUFFICIENT_BUFFER) {
+            buf.resize(size);
+            continue;
+        }
+        return;
+    }
+
+    auto* table = reinterpret_cast<MIB_UDPTABLE_OWNER_PID*>(buf.data());
+    for (DWORD i = 0; i < table->dwNumEntries; ++i) {
+        auto& row = table->table[i];
+        auto& proc = lookup_proc(row.dwOwningPid, cache);
+        ctx.write_output(std::format("udp|{}|{}|*|0||{}|{}|{}", format_addr4(row.dwLocalAddr),
+                                     ntohs(static_cast<u_short>(row.dwLocalPort)),
+                                     static_cast<int>(row.dwOwningPid), escape_pipes(proc.name),
+                                     escape_pipes(proc.path)));
+    }
+}
+
+void emit_udp6_attributed(yuzu::CommandContext& ctx, std::unordered_map<DWORD, ProcInfo>& cache) {
+    DWORD size = 0;
+    GetExtendedUdpTable(nullptr, &size, FALSE, AF_INET6, UDP_TABLE_OWNER_PID, 0);
+    if (size == 0)
+        return;
+
+    std::vector<BYTE> buf(size);
+    for (int attempt = 0; attempt < 3; ++attempt) {
+        DWORD ret = GetExtendedUdpTable(buf.data(), &size, FALSE, AF_INET6, UDP_TABLE_OWNER_PID, 0);
+        if (ret == NO_ERROR)
+            break;
+        if (ret == ERROR_INSUFFICIENT_BUFFER) {
+            buf.resize(size);
+            continue;
+        }
+        return;
+    }
+
+    auto* table = reinterpret_cast<MIB_UDP6TABLE_OWNER_PID*>(buf.data());
+    for (DWORD i = 0; i < table->dwNumEntries; ++i) {
+        auto& row = table->table[i];
+        auto& proc = lookup_proc(row.dwOwningPid, cache);
+        ctx.write_output(std::format("udp6|{}|{}|*|0||{}|{}|{}", format_addr6(row.ucLocalAddr),
+                                     ntohs(static_cast<u_short>(row.dwLocalPort)),
+                                     static_cast<int>(row.dwOwningPid), escape_pipes(proc.name),
+                                     escape_pipes(proc.path)));
+    }
+}
+
+void enumerate_and_stream_attribution(yuzu::CommandContext& ctx) {
+    std::unordered_map<DWORD, ProcInfo> cache;
+    emit_tcp4_attributed(ctx, cache);
+    emit_tcp6_attributed(ctx, cache);
+    emit_udp4_attributed(ctx, cache);
+    emit_udp6_attributed(ctx, cache);
+}
+
 #endif // platform
 
 // ── ABI4 capability declarations (#2204) ────────────────────────────────────
 //
-// One action, native in-process on every platform: Linux reads
-// /proc/net/{tcp,tcp6,udp,udp6} + /proc/[pid]/fd directly, macOS uses
-// libproc, Windows uses the IP Helper API — zero subprocesses anywhere
-// (rung 1).
+// Two actions, both native in-process on every platform: Linux reads
+// /proc/net/{tcp,tcp6,udp,udp6} + /proc/[pid]/fd directly (attribution also
+// /proc/[pid]/{comm,exe}), macOS uses libproc via the shared
+// macos_socket_walk.hpp, Windows uses the IP Helper API (attribution also
+// QueryFullProcessImageNameW) — zero subprocesses anywhere (rung 1).
 const YuzuActionDescriptor kActionDescriptors[] = {
     {
         /* .action      = */ "netstat_list",
@@ -558,6 +784,14 @@ const YuzuActionDescriptor kActionDescriptors[] = {
         /* .macos_leg   = */ {YUZU_SUPPORT_SUPPORTED, 1, "libproc", nullptr},
         /* .windows_leg = */
         {YUZU_SUPPORT_SUPPORTED, 1, "GetExtendedTcpTable/GetExtendedUdpTable", nullptr},
+    },
+    {
+        /* .action      = */ "attribution",
+        /* .linux_leg   = */
+        {YUZU_SUPPORT_SUPPORTED, 1, "/proc/net/* + /proc/[pid]/{comm,exe,fd}", nullptr},
+        /* .macos_leg   = */ {YUZU_SUPPORT_SUPPORTED, 1, "libproc", nullptr},
+        /* .windows_leg = */
+        {YUZU_SUPPORT_SUPPORTED, 1, "IP Helper API + QueryFullProcessImageNameW", nullptr},
     },
 };
 
@@ -572,7 +806,7 @@ public:
     }
 
     const char* const* actions() const noexcept override {
-        static const char* acts[] = {"netstat_list", nullptr};
+        static const char* acts[] = {"netstat_list", "attribution", nullptr};
         return acts;
     }
 
@@ -591,6 +825,9 @@ public:
         if (action == "netstat_list") {
             return do_list(ctx);
         }
+        if (action == "attribution") {
+            return do_attribution(ctx);
+        }
         ctx.write_output(std::format("unknown action: {}", action));
         return 1;
     }
@@ -599,6 +836,16 @@ private:
     int do_list(yuzu::CommandContext& ctx) {
 #if defined(__linux__) || defined(__APPLE__) || defined(_WIN32)
         enumerate_and_stream(ctx);
+        return 0;
+#else
+        ctx.write_output("error: network enumeration not supported on this platform");
+        return 1;
+#endif
+    }
+
+    int do_attribution(yuzu::CommandContext& ctx) {
+#if defined(__linux__) || defined(__APPLE__) || defined(_WIN32)
+        enumerate_and_stream_attribution(ctx);
         return 0;
 #else
         ctx.write_output("error: network enumeration not supported on this platform");
