+++ agents/shared/user_profile_model.hpp
@@ -216,7 +216,7 @@ struct ProfileInfo {
 /// one row shape, and the dashboard's split_fields (result_parsing.hpp) has
 /// no per-plugin entry for "registry" so it falls to the generic
 /// split-on-all-pipes path -- every other multi-column plugin (procfetch,
-/// netstat, sockwho, vuln_scan) emits raw pipe-joined field values with no
+/// netstat, vuln_scan) emits raw pipe-joined field values with no
 /// self-describing prefix for exactly this reason. A leading "profile|" tag
 /// was an extra field split_fields didn't know to skip, shifting every real
 /// column one position right of its header.
