+++ changelog.d/3403-netstat-sockwho-consolidation.changed.md
@@ -1,0 +1,1 @@
+- **`netstat` gains a new `attribution` action; the `sockwho` plugin is retired.** `attribution` enumerates the same TCP/UDP connections and listening sockets as `netstat_list`, additionally resolving each socket's owning process name and executable path — the functionality the standalone `sockwho` plugin previously provided. `sockwho` and its `sockwho_list` action are removed; existing content built on `device.network.sockwho_list` should move to the new `device.network.netstat_attribution` definition.
