+++ content/definitions/netstat_attribution.yaml
@@ -1,27 +1,31 @@
-## Socket-to-Application Mapping Definitions
-## Maps open sockets to owning processes with full path resolution.
-## Plugin: sockwho | Actions: sockwho_list
+## Netstat Attribution Definitions
+## Enumerates active network connections and listening sockets, resolving
+## the owning process for each. Folds the retired `sockwho` plugin's
+## functionality into netstat as a new `attribution` action (#3403).
+## Plugin: netstat | Actions: attribution
 ##
 ## Sources by platform:
 ##   Windows: IP Helper API + QueryFullProcessImageNameW for path resolution
 ##   Linux:   /proc/net/* + /proc/[pid]/{comm,exe,fd} for inode-to-PID mapping
 ##   macOS:   libproc (proc_listallpids, proc_name, proc_pidpath, proc_pidfdinfo)
+##            via the shared agents/shared/macos_socket_walk.hpp helper
 
 ---
 apiVersion: yuzu.io/v1alpha1
 kind: InstructionDefinition
 metadata:
-  id: device.network.sockwho_list
-  displayName: Socket-to-Application Map
+  id: device.network.netstat_attribution
+  displayName: Network Connections with Process Attribution
   version: 1.0.0
   description: >
-    Enumerates all open TCP/UDP sockets and resolves the owning process
-    for each, returning PID, process name, executable path, and full
-    socket details (protocol, local/remote addresses, ports, state).
-    On Windows uses the Extended TCP/UDP Table APIs and resolves
-    process paths via QueryFullProcessImageNameW with a per-PID cache.
-    On Linux scans /proc to build inode-to-PID and PID-to-process maps
-    in a single pass. On macOS uses libproc for per-process FD
+    Enumerates all active TCP/UDP connections and listening sockets on
+    the endpoint, resolving the owning process for each and returning
+    protocol, local and remote addresses and ports, connection state,
+    owning PID, process name, and executable path. Covers TCP4, TCP6,
+    UDP4, and UDP6. On Windows uses the Extended TCP/UDP Table APIs and
+    resolves process paths via QueryFullProcessImageNameW with a per-PID
+    cache. On Linux scans /proc to build inode-to-PID and PID-to-process
+    maps in a single pass. On macOS uses libproc for per-process FD
     enumeration with proc_name and proc_pidpath. Pipe characters in
     process names and paths are escaped. Useful for security
     investigations to identify which application owns a suspicious
@@ -33,8 +37,8 @@ spec:
   platforms: [windows, linux, darwin]
 
   execution:
-    plugin: sockwho
-    action: sockwho_list
+    plugin: netstat
+    action: attribution
     concurrency: per-device
 
   parameters:
@@ -43,12 +47,6 @@ spec:
 
   result:
     columns:
-      - name: pid
-        type: int32
-      - name: process_name
-        type: string
-      - name: process_path
-        type: string
       - name: proto
         type: string
       - name: local_addr
@@ -61,6 +59,12 @@ spec:
         type: int32
       - name: state
         type: string
+      - name: pid
+        type: int32
+      - name: process_name
+        type: string
+      - name: process_path
+        type: string
     aggregation:
       groupBy:
         - process_name
@@ -68,7 +72,7 @@ spec:
       operations:
         - count
 
-  readablePayload: "Map sockets to owning applications"
+  readablePayload: "List network connections with owning process"
 
   gather:
     ttlSeconds: 30
@@ -82,7 +86,7 @@ spec:
 
   compatibility:
     minAgentVersion: 1.0.0
-    requiredPlugins: [sockwho]
+    requiredPlugins: [netstat]
 
 status:
   phase: active
