+++ docs/agent-privilege-model.md
@@ -92,7 +92,7 @@ This table is the source of truth for what each plugin requires. When you add a 
 | `services.list`, `services.running` | default | default | default |
 | `certificates.list`, `certificates.details` | System / root keychains: default (`security find-certificate`). **The console user's login keychain (`store=login`, and `store=all` when a user is present) requires root:** the shipped LaunchDaemon (no `UserName` — runs as root, see the `tar.*` row / TL;DR) runs `launchctl asuser <uid> sudo -n -u <user> /usr/bin/security find-certificate`, hopping into that user's per-session launchd/Aqua (GUI) context to reach their login keychain. `launchctl asuser` itself requires root, so this path is live only because the daemon runs as root today; there is no reachable login keychain if nobody is logged in at the console. When #1455 narrows the agent to the non-root `_yuzu` account, this path needs the dormant sudoers grant documented in `agents/shared/macos_console_user.hpp` (`_yuzu ALL=(root) NOPASSWD: /bin/launchctl asuser * /usr/bin/sudo -n -u * /usr/bin/security find-certificate -a -p *`) — added with that PR, not the default install today. | default | default |
 | `tar.*` (TAR snapshots) | **root** (LaunchDaemon, no `UserName` — see TL;DR). The Endpoint Security process stream's `es_new_client` requires root + the `com.apple.developer.endpoint-security.client` entitlement; without them it falls back to the `KERN_PROC_ALL` sysctl poll. | `cap_dac_read_search` (read all `/proc/<pid>/*`) | `SeBackupPrivilege` (read regardless of DACL). **Process capture runs a real-time ETW session on `Microsoft-Windows-Kernel-Process`** — enabled by the agent account's `Performance Log Users` membership (see Group memberships above; no extra LSA privilege). Per-process user attribution uses `OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION)` on still-alive processes only. **The `software` source** reads the machine-wide HKLM Uninstall keys only (machine scope; no per-user hives, no `NTUSER.DAT` mounting), so it needs no privilege beyond the agent account's existing registry-read access. **The mapdrive outbound-history collector** (`enum_registry_outbound_history`, #2771) walks each profile's `HKCU\Network\<Drive>` history the same way `installed_apps.list_per_user` does — loaded hives need no privilege; an offline mount additionally rides `SeRestorePrivilege` (already held, same grant as `SeBackupPrivilege` above). See the deployment-footprint note below for the boot AutoLogger. |
-| `netstat.*`, `sockwho.*` | default | default | default |
+| `netstat.*` | default | default | default |
 | `antivirus.*` | default — `PlistBuddy`/`stat` reads of the XProtect/MRT bundles, `systemextensionsctl list`, and `pgrep` are all unprivileged | default (`pgrep`, directory existence checks) | default (PowerShell SecurityCenter2 / `Get-MpComputerStatus` CIM queries) |
 | `firewall.state`, `firewall.rules` | default for the primary Application Firewall read (`socketfilterfw --getglobalstate` is unprivileged). The secondary pf read (`pfctl -s info` / `-s rules`) opens `/dev/pf` and needs **root** — the LaunchDaemon runs as root today (see TL;DR); an unprivileged agent reports `pf\|unknown` (state) / no pf rows (rules), never a false-safe value. No firewall-specific sudoers entry — the plugin never mutates firewall state and deliberately does not ride the quarantine plugin's `/sbin/pfctl` grant (its pfctl calls are read-only, unprefixed by sudo), so under the least-privilege model the pf row degrades to `unknown`. | `firewall-cmd --state` is an unprivileged D-Bus query; the `ufw status` / `iptables -L` fallbacks need root, so on a non-firewalld host the unprivileged agent can misreport `backend\|none` (pre-existing limitation, tracked for the Linux parity sweep) | default — `netsh advfirewall show` queries need no elevation |
 | `bitlocker.state` | n/a | n/a | `SeBackupPrivilege` (some BitLocker WMI properties require this on system drives) |
